src/send_module/IM_bot/pyproject.toml
[tool.poetry]
name = "im_bot"
version = "0.1.0"
description = "IM all in one"
authors = ["XDeviation <191602098@qq.com>"]
readme = "README.md"
packages = [{include = "im_bot"}]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"
requests = "^2.28.1"
wheel = "^0.38.4"
sqlalchemy = "^1.4.43"
psycopg2 = "^2.9.5"
backports-zoneinfo = "^0.2.1"
logger = {path = "../../logger/dist/logger-1.1.3-py3-none-any.whl"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
